feat(client): adding function for  checking user dependencis function
